fix: rename tilia.logging
This avoids shadowing the standard library `logging` module in some cases
