+++ tilia/file/autosave.py
@@ -7,7 +7,7 @@ from typing import Callable
 
 import tilia.constants
 from tilia import dirs
-from tilia.logging import logger
+from tilia.log import logger
 from tilia.settings import settings
 from .common import are_tilia_data_equal, write_tilia_file_to_disk
 from .tilia_file import TiliaFile
